sawp-backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
docopt==0.6.2
idna==2.10
Js2Py==0.71
packaging==21.0
psycopg2==2.9.1
pycparser==2.20
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
